Time: N/A (0%), Space: N/A (0%) - LeetHub

+++ 0098-validate-binary-search-tree/0098-validate-binary-search-tree.cpp
@@ -11,12 +11,12 @@
  */
 class Solution {
 public:
-    bool isValid(TreeNode*root, long minVal, long maxVal){
+    bool isValid(TreeNode*root,int minVal,int maxVal){
         if(!root)return true;
-        if(root->val<=minVal || root->val>=maxVal)return false;
+        if(root->val<=minVal||root->val>=maxVal)return false;
         return isValid(root->left,minVal,root->val)&&isValid(root->right,root->val,maxVal);
     }
     bool isValidBST(TreeNode* root) {
-        return isValid(root,LONG_MIN,LONG_MAX);
+        return isValid(root,INT_MIN,INT_MAX);
     }
 };
